Website design change
The site's design has been changed to be similar to a past Power BI project.

It takes on a more space-tech-orientated feel for 2025.

The 2024 version of the site was too colourful, which distracted attention away from the content.

This more simple and monotone design should help the content stand out better.